🐛 fix(nextjs): downgrade next.js version to 14.0.1
The changes in this commit fix an issue with the `next` package version across the `packages/auth`, `tooling/eslint-config`, and `apps/nextjs` packages. The version of `next` is downgraded from `14.0.3` to `14.0.1` to address potential compatibility issues.

This change is necessary to ensure that the application can be built and run correctly across the different packages and environments.

🌟feat(UI): add some intersting component