Admin dashboard + refactor UI  (#6)
* add admin dashboard

* 🐛 fix(nextjs): downgrade next.js version to 14.0.1

The changes in this commit fix an issue with the `next` package version across the `packages/auth`, `tooling/eslint-config`, and `apps/nextjs` packages. The version of `next` is downgraded from `14.0.3` to `14.0.1` to address potential compatibility issues.

This change is necessary to ensure that the application can be built and run correctly across the different packages and environments.

🌟feat(UI): add some intersting component

* ✨ feat(middleware): Improve locale handling and redirect logic

The changes made in this commit improve the handling of locales and the redirect logic in the Next.js middleware. The key changes are:

- Removed the `redirect` import from the top of the file, as it is now imported from `next/navigation`.
- Simplified the `getLocale` function by removing the need to transform the request headers into a plain object.
- Removed the `withAuth` import from `next-auth/middleware`, as it is not being used in this file.
- Refactored the `isNoRedirect` and `isPublicPage` functions to use more concise regular expressions.
- Simplified the `isNoNeedProcess` function to use a more concise regular expression.
- Streamlined the middleware function by removing unnecessary checks and simplifying the redirect logic.

These changes improve the overall readability and maintainability of the middleware file, while also ensuring that the locale handling and redirect logic are more robust and efficient.